// src/config/enum.ts
export enum Adapter {
  Spotify = "spotify",
  YouTube = "youTube",
  AppleMusic = "appleMusic",
  Deezer = "deezer",
  SoundCloud = "soundCloud",
  Tidal = "tidal",
}

export enum Parser {
  Spotify = "spotify",
  YouTube = "youTube",
  AppleMusic = "appleMusic",
  Deezer = "deezer",
  SoundCloud = "soundCloud",
}

export enum MetadataType {
  Song = "song",
  Album = "album",
  Playlist = "playlist",
  Artist = "artist",
  Podcast = "podcast",
  Show = "show",
}